convert tabs to spaces to fix alignment of parameters and comments
When I wrote this seven years ago, I wrote it in an LC3 editor that
defaulted to 6-character tab widths. I used tabs pretty heavily to
align parameters and comments, and although that made editing convenient,
viewing the file in various editors is pretty painful now.

Rather than trying to find a way to convince numerous editors to use
6-character tab widths, and rather than replacing only the alignment
tabs with spaces, I figure the easiest solution is to just move to
spaces throughout the entire file.